Add Room
Our rooms are network enabled. They will be the core
godot nodes involved in the server (beside the server
itself). We are relying on NodePaths for our RPC calls,
the idea is that we have a root network node (our client/server
connections) and under those we have our room network nodes
that handle calls to specific clients as well as distributing
calls to the rest of the game.