Add Basic Game Initilization (#2)
- Add Empty Constructor To Match
- Add Zone Lists To Player
- Add Card Data (with [Tool])
- Add Tests
- Add Load Deck Function
- Add Draw Starting Hand Function
- Add Draw Function To Player & Match
- Refactor Player Queuing


Notes:

We are required to add an empty constructor to any Class that inherits from a Godot Object if we have given it a non-empty constructor. This is likely required by Godot behind the scenes.

Card Data must use the [Tool] keyword if we want to run our tests in "launch-in-editor" context.

Previously the server registered the player but now that we are interested in their deck list contents, we inverted this and had the player send their deck list over once they are connected.

We have two primary scripts that invoke actions. The Rules-enforced type (namely Match.cs) which will make sure the play is legal before committing to it and the Actor type script (player or cards) that can invoke actions without caring about the rules. This allows us to use the player classes to invoke actions directly in tests without worrying about setting it up. Whereas we can have longer and more robust tests with the Rules-enforced type.