Add Card Register
Card Register is a list of cards in the game that will maintain
order. When we want to load player decks, we pass them a copy of the
card register. This allows us to avoid using the network based Room
Script and instead deal with Match.cs directly which should improve
testing.

When we add a card to the register, our card id will be equal to the
register count BEFORE adding it, which means its element index should
be equal to its Id.