Refactor Match Init
Moved a lot of game initialization work from Room to Match.
This mainly allows us to reduce assignments in our tests which
helps prevent flaky tests. This init doesn't require waiting for players
either because once the room is created, everything is ready before we
get into our ready call, which when we do simply calls Update() to
send all of this data to players instead of calling an unnecessary
start function on match.

With that in mind too, in previous incarnations of the game we used
to send every event seperately to the user but maybe we could alternatively
queue everything up locally and then only send it once the Update() method
is called.