Add Player Actions (#5)
* Add SetFaceDown Function
* Add DeclareAttack Function
* Add DirectAttack Function Rules
* Add Happy Path Match Tests

Notes:

The deploy function was added a bit too quickly
in our previous branch so we're just ticking it off here.

We're leaving the actual implementation of battle & activation until
the next stage(s). We just added the basic rules here.

We added a number of tests as we do with all features but now we added the happy path tests as well which allows
us to cross-reference actions in match between the disqualification actions in our rules tests.