Add Base(Server)Test
Basic Custom Test for our server tests. It handles basic initialization.
We can be sure it won't ever get run itself as long as its class and file names
do not match. We have no need to modify the decklist because we can just modify
the cards on demand for our tests which exist in a controlled enviroment.