Add Skill Objects
- Newtonsoft JSON Skill Initialization
- SkillInfo
- SkillObject
- Skill Trigger/Instruction Enumerations

Notes:
Even if a card doesn't have a skill, we still need to give
them the skill property in JSON to avoid null reference.

Skills hold a readonly copy of instructions but then also added
an argument lists. Unfortunately we have to cast everything here
to an object (unless we add an argument wrapper? but that's too
in the future to be concerned about).