Add Set Card Skills
- Change Title / Faction / Power

Changing CardType is possibly but may not be wise
to do it immediatly.

We push enum values to arguments as Strings (this way they're
more meaningful in JSON files and less prone to error).

When affecting cards, we always add them to a group. This is handy
when we want to add multiple targets together but not entire zones.