Add Conditional Instructions
- IfLessThanInstruction
- Jump Instruction
- GoTo End Instruction
- Discard Argument Instruction

We require discard argument instruction in case we jump
to an else branch and suddenly stop caring about some arguments.

We need a few more conditionals (mainly equal / not equal and
greater than). The jury is still out on the IfEqualOR[Greater/Lesser]
than since we likely will be able to do most of the work with just
If/Less.

With this in place, our main concerns are now Targeting / Triggers
and Constants.