Streamline Stack (#10)
* Remove Argument Array

We flat-lined our stack machine to a single
set of int instructions. We're using our custom list
other than CSharps built-in stack which for some reason
adds elements to the beginning of the list.

This also means we removed our change title power. Besides
it being difficult to work with the int list, it would also
probably get confusing in different languages so better to just
add that kind of information to the cards themselves via Faction
or some such.

* Remove All References To Argument List

No longer needed.

* Remove Dead Code From Virtual Stack Machine

Largely redundant casts.