Refactor GetCard(Func<Player>) To GetCards(GetPlayer().<Zone>
There was no reason for this to be a confusing function. We
grab the player, get the property and from there have getcards
consume it (into _cards.AddRange). Since the actual
movement is based on the card zone property, it doesn't matter if
the list we pass is a direct reference or not.