Update GUI Display. Add Organization Nodes.
Rather than have the cards be messy around with the Z-Index, we
are going to add them to base nodes at the very end of the tree so
they're over everything.

(In retrospect, we don't even technically need any of the containers
beyond for UI reasons if the cards are not in them BUT we could use them
as guides if we don't get rid of them entirely).