Merge CommandQueue Into Room; Pass Room to Commands
The command queue didn't have much reason to exist thanks to
our loop for generating a command dictionary. So we merged it back in.

We also now pass our room to commands to trigger the relevant effects
since it already has the ability to reach everywhere. This lets our
command objects do the heavy lifting.